requirement.txt
opencv-python
numpy
pillow
dlib
tkinter